(#24) Add more BuildProviders
This commit adds in additional BuildProviders to allow the recipe to be
ran on TeamCity and GitHub Actions, as well as when running locally on a
developer machine. This has been borrowed from other recipes, including
Cake.Recipe and Chocolatey.Cake.Recipe.  The existing AppVeyor Build
Provider functionality has also been modified to fit with this concept,
as well as altering the DAG of the tasks to ensure correct execution.