(#10) Allow usage of .NET Global Tools for tools
Similar to how Cake.Recipe does things, a new BuildParameter has been
introduced called preferDotNetGlobalToolUsage, which when set, will
prefer to use .NET Global Tool variants of the tools that are in play,
for example, GitVersion and GitReleaseManager. To make this possible, a
new class called ToolSettings has been introduced, which allows for
control over what versions for these tools are actually used.  This
allows for direct control when a new version of a given tool is
available, that hasn't been updated in the recipe yet.